+++ src/constants.ts
@@ -1,7 +1,8 @@
 import { TMat2D } from './typedefs';
+// use this syntax so babel plugin see this import here
+import { version } from '../package.json';
 
-// TODO: consider using https://github.com/swiing/rollup-plugin-import-assertions so we can import json in node and have rollup build pass
-export { version as VERSION } from '../package.json';
+export const VERSION = version;
 export function noop() {}
 export const halfPI = Math.PI / 2;
 export const twoMathPi = Math.PI * 2;